pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "spanish-eit-scorer"
version         = "1.0.0"
description     = "Deterministic, rubric-driven automated scoring for the Spanish Elicited Imitation Task"
readme          = "README.md"
requires-python = ">=3.9"
license         = {text = "MIT"}
keywords        = ["NLP", "SLA", "Spanish", "EIT", "automated-scoring", "reproducible"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Intended Audience :: Science/Research",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "tqdm>=4.65",
    "spacy>=3.5",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "black>=23.0",
]

[project.scripts]
eit-score = "eit_scorer.cli:main"
eit-api   = "apps.api.main:run"
eit-gen   = "eit_scorer.synthetic.generate:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["eit_scorer*", "apps*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.ruff]
line-length = 100
select      = ["E", "F", "W", "I"]
